debounce the user input
used the setTimeout and clearTimeout (inside the cleanup func) to delay checking the validity of the enteredEmail and the enteredPassword for 500ms after the user has stopped entering anything.
The cleanup func runs before the effect func runs (everytime, except for the first time), and since it runs before the effect func, the last timer is not cleared.